+++ simulator/placing_strategy/partition_policy/UsedFirst.h
@@ -6,7 +6,7 @@
 namespace placing_strategy::partition_policy {
 
 struct UsedFirst {
-    static unsigned Partition(std::vector<const Machine*>& machines);
+    static size_t Partition(std::vector<const Machine*>& machines);
 };
 
 } // namespace placing_strategy::partition_policy
